Add cowpacman2srm script
Create SystemRescueModules (SRM) from pacman packages installed into the COW space

This script is meant to help creating SystemRescueModules (SRM) for SystemRescue.
More info about SRMs and this script can be found at:
https://www.system-rescue.org/Modules/

To use first install all packages you want to have in your SRM with pacman.
Default COW (Copy-On-Write) space is a ramdisk, so you usually don't have to do anything
special except provide enough RAM.
You can also enable any systemd services that are in these packages.

Then call:
cowpacman2srm [-c compalg] [-l complevel] targetfile.srm

Copy the .srm file to archisobasedir (default: "sysresccd") on your boot disk
and add the "loadsrm" boot parameter to SystemRescue.

There is no mechanism to check if a SRM is compatible with the version of SystemRescue
you are trying to run it with. So it is higly recommended to only use this script on
the exact version of SystemRescue you plan to use the SRM with.