Support chrooting on older systems without usrmerge
On newer systems /bin is linked to /usr/bin, the same goes for /sbin.
This is called usrmerge. Some distributions seem to have introduced
usrmerge just very recently, like openSUSE in 2021. So SystemRescue
should fully support this for years to come.

So we need /bin, /sbin and /usr/sbin in the default $PATH to make it
easy to use from within a chroot.