explicitly list some wanted packages that are currently installed through dependencies or group
Recently SystemRescue had two regressions due to packages going missing silently (inetutils and net-tools).
This happened because these packages where pulled in through groups or dependencies and Arch decided to
either reduce the dependencies or drop them from the group. This might happen again in the future.

To reduce the possibility of this, explicitly add packages to the package list that contain tools deemed
wanted and are currently only pulled in through dependency or group. Which packages are deemed
wanted is obviously a bit subjective and could be discussed or changed if necessary.

This commit doesn't add any new packages that weren't already in SystemRescue.