add "mountall" command
It loops through all available block devices in the system and tries to mount them.
They are mounted to /mnt/<devicename>.

If they have "dev", "proc" or "sys" dirs in them, then the respective dirs of the
running SystemRescue are bind-mounted into it. This allows running some commands
like grub-install when chrooting into the dir.