Add workaround for systemd not re-evaluating dependencies while running a transaction
As described in #279, systemd does not re-evaluate the dependencies of changed units and
targets when daemon-reload is executed. It just continues it's transaction as it was
configured when starting the transaction.

So a late-loaded SRM can't autostart a service.

To work around this, add an extra call to "systemctl --no-block start multi-user.target"
after the daemon-reload. This will create a new transaction after the current one and
allow new services to be added to multi-user.target in a late-loaded SRM.

This means of course that this workaround only works for multi-user.target, not other
services or targets.