fix dns requests in sysrescue-configuration.lua (#291)
The dns library used by lua requires /etc/hosts to exist, otherwise all
requests for name resolution are denied. This meant that using the
sysrescuecfg boot option with a http URL and hostname did not work
while using an IP worked.

This patch adds an empty /etc/hosts file to the initramfs image to fix this.