add sysresccd/.imageinfo file to the iso image, allows systemrescue-usbwriter to determine compatibility
The upcoming systemrescue-usbwriter should work mostly independent of SystemRescue .iso version
and one systemrescue-usbwriter version should work for a wide range of SystemRescue versions.
For this to work, it has to exactly determine compatibility without any crude guesswork.
The necessary info for this is stored in the imageinfo file.

It also contains a "FORMAT_EPOCH" field that allows to explicitly declare incompatibility to older
versions of systemrescue-usbwriter in case automatic compatibility detection based on syslinux
version etc. doesn't work.