reduce timeout to automatic boot of SystemRescue with default settings from 90 seconds to 30
Consider you don't have a keyboard connected and can only connect over the network. Or the
keyboard you use doesn't work in the bootloader due to BIOS issues with USB. Then waiting
the whole 90 seconds is cumbersome.

30 seconds should still be enough to see the bootloader over a slow network KVM, get the
correct monitor or keyboard attached and similar things. So it should still be enough time
to modify the default boot if desired.