add customization hooks to boot menus
Adds empty sysresccd_custom.cfg and /boot/grub/custom.cfg files that
are included from the regular boot menus (syslinux and grub). These
are intended for the user to customize.

Having dedicated files for this allows completely replacing the files.
This is easier than requiring careful patching with potential conflicts
with different versions of SystemRescue.